requirements.txt
google-generativeai==0.8.4
Pillow>=10.0.0
firebase-admin>=6.0.0
python-dotenv>=1.0.0
google-cloud-firestore>=2.11.0
flask==3.0.2
flask-cors==4.0.0
